Searchbox, first pass
