+++ packages/office-ui-fabric-react/src/components/SearchBox/SearchBox.types.ts
@@ -1,4 +1,6 @@
 import * as React from 'react';
+import { ITheme, IStyle } from '../../Styling';
+import { IStyleFunction } from '../../Utilities';
 
 export interface ISearchBox {
   /**
@@ -67,4 +69,23 @@ export interface ISearchBoxProps extends React.InputHTMLAttributes<HTMLInputElem
    * @default false
    */
   underlined?: boolean;
+
+  theme?: ITheme;
+
+  getStyles?: IStyleFunction<ISearchBoxProps, ISearchBoxStyles>
+}
+
+export interface ISearchBoxStyleProps {
+  theme: ITheme;
+  disabled?: boolean;
+  hasFocus?: boolean;
+  underlined?: boolean;
+}
+
+export interface ISearchBoxStyles {
+  root?: IStyle;
+  iconContainer?: IStyle;
+  icon?: IStyle;
+  field?: IStyle;
+  clearButton?: IStyle;
 }
